[hlsl] Fix DecomposeUniformAccess to cast index to u32 for LoadVectorElement
Bug: 42251045
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/203655
Commit-Queue: Antonio Maiorano <amaiorano@google.com>

+++ src/tint/lang/hlsl/writer/access_test.cc
@@ -979,7 +979,7 @@ void foo() {
   uint x = 1u;
   vector<float16_t, 4> a = tint_bitcast_to_f16(v[0u]);
   float16_t b = float16_t(f16tof32(v[0u].x));
-  uint v_4 = (min(x, 3u) * 2u);
+  uint v_4 = (uint(min(x, 3u)) * 2u);
   uint v_5 = v[(v_4 / 16u)][((v_4 % 16u) / 4u)];
   float16_t c = float16_t(f16tof32((v_5 >> ((((v_4 % 4u) == 0u)) ? (0u) : (16u)))));
   float16_t d = float16_t(f16tof32(v[0u].y));
